Projet/requirements.txt
Flask
azure-storage-blob
azure-ai-textanalytics
azure-core